Reorganize files, move SearchBar to its own component file
