+++ apps/frontend/components/flows/AddExperiment/stepComponents/ParamStep.tsx
@@ -1,6 +1,6 @@
 import { Fragment } from 'react';
 import { DragDropContext, Droppable } from 'react-beautiful-dnd';
-import { InputSection } from '../InputSection';
+import { InputSection } from '../../../InputSection';
 
 export const ParameterOptions = ['integer', 'float', 'bool', 'string'] as const;
